refactor: move dispatcher servlet to mvc
